[sival,pwrmgr] Move pwrmgr sleep all resets tests out sim_dv
These tests can be run in cw310, so they don't belong in sim_dv.

Addresses #20135
